Move and add docs
